Check for serialized data to allow faster search-replace
Adds a check to see if serialized data is present and if not runs a
direct SQL query to search-replace the database. Also introduces a
--safe flag to force use of the serialize-safe serach-replace method
for every row. Adds a check to see if the string lengths are the same
which will trigger the direct sql queries for all search-replace since
this will not invalidate serialized data. Finally, refines existing
behat tests and adds a new scenario to ensure proper function of the
new fast replace method and flags.